% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo.r
\name{anti.trafo}
\alias{anti.trafo}
\title{Inverse log-transform of child ages}
\usage{
anti.trafo(x, adult.age = 20)
}
\arguments{
\item{x}{Age estimate where child ages have been log-transformed.}

\item{adult.age}{(Default: 20)}
}
\value{
Transformed age estimates transformed back to years.
}
\description{
Inverse log-transform of child ages
}
