% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accel.r
\name{meffonym.accel}
\alias{meffonym.accel}
\title{Calculate age acceleration}
\usage{
meffonym.accel(
  x,
  model,
  age,
  calibrate = F,
  transform = ifelse(model \%in\% c("horvath", "pedbe"), anti.trafo, NULL),
  adjust = NULL
)
}
\arguments{
\item{x, model, calibrate, transform, adjust}{See \code{\link{meffonym.score}}.}

\item{age}{Chronological age in years.}
}
\value{
Age-adjusted model scores.
}
\description{
Calculate age acceleration
}
