% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.r
\name{meffonym.add.model}
\alias{meffonym.add.model}
\title{Add or update set of available models}
\usage{
meffonym.add.model(name, variables, coefficients, description, intercept = 0)
}
\arguments{
\item{name}{Name of the model.}

\item{variables}{Names of CpG sites included in the model.}

\item{coefficients}{Coefficients of the CpG sites in the model.}

\item{description}{Text description of the model, e.g. publication.}

\item{intercept}{Model intercept (Default: 0).}
}
\description{
Add or update set of available models
}
