% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmiq.r
\name{meffonym.bmiq.calibration}
\alias{meffonym.bmiq.calibration}
\title{BMIQ normalization to a standard}
\usage{
meffonym.bmiq.calibration(x, standard, ...)
}
\arguments{
\item{x}{DNA methylation matrix (rows=CpG sites, columns=samples).}

\item{standard}{A named vector of methylation levels corresponding to the row names of \code{x}.}

\item{...}{See \url{http://code.google.com/p/bmiq/} for more details.}
}
\value{
Matrix identical to \code{x} but with methylation levels normalized
by BMIQ to the \code{standard} and rows corresponding to \code{standard}.
Missing values in \code{x} are imputed by k-nearest neighbor.
}
\description{
Methylation levels are normalized so that they are comparable to a given dataset. It is based on
BMIQ:
}
\details{
A beta-mixture quantile normalization method for correcting probe design bias in Illumina Infinium 450 k DNA methylation data.
Teschendorff AE, Marabita F, Lechner M, Bartlett T, Tegner J, Gomez-Cabrero D, Beck S.
Bioinformatics. 2013 Jan 15;29(2):189-96. doi: 10.1093/bioinformatics/bts680. Epub 2012 Nov 21.
PMID: 23175756

The code was obtained from \url{http://labs.genetics.ucla.edu/horvath/dnamage/NORMALIZATION.R}.
}
