% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunedinpace.r
\name{meffonym.dunedinpace.estimate}
\alias{meffonym.dunedinpace.estimate}
\title{DunedinPACE pace of aging estimate}
\usage{
meffonym.dunedinpace.estimate(x)
}
\arguments{
\item{x}{DNA methylation matrix (rows=CpG sites, columns=samples).}
}
\value{
Scores derived for each sample in \code{x}.
}
\description{
DunedinPACE pace of aging estimate
}
