% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunedinpace.r
\name{meffonym.dunedinpace.standard}
\alias{meffonym.dunedinpace.standard}
\title{DunedinPACE methylation standard}
\usage{
meffonym.dunedinpace.standard()
}
\description{
Obtained from
\url{https://github.com/danbelsky/DunedinPACE}.
}
