% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epismoker.r
\name{meffonym.epismoker}
\alias{meffonym.epismoker}
\title{Epismoker smoking status probabilities}
\usage{
meffonym.epismoker(x, female, calibrate = F)
}
\arguments{
\item{x}{DNA methylation matrix (rows=CpG sites, columns=samples).}

\item{female}{Vector of booleans indicating whether each individual is female.}
}
\value{
Matrix of epismoker smoking status probabilities (rows=samples, columns=smoking status)
}
\description{
Epismoker smoking status probabilities
}
