% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epismoker.r
\name{meffonym.epismoker.score}
\alias{meffonym.epismoker.score}
\title{Epismoker linear predictor for a specified smoking status}
\usage{
meffonym.epismoker.score(x, group, female, calibrate = F)
}
\arguments{
\item{x}{DNA methylation matrix (rows=CpG sites, columns=samples).}

\item{group}{Either 'current', 'former' or 'never'}

\item{female}{Vector of booleans indicating whether each individual is female.}
}
\value{
Epismoker smoking group linear predictor derived for each sample in \code{x}.
}
\description{
Epismoker linear predictor for a specified smoking status
}
