% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horvath.r
\name{meffonym.horvath.age}
\alias{meffonym.horvath.age}
\title{Horvath DNA methylation age estimate}
\usage{
meffonym.horvath.age(x)
}
\arguments{
\item{x}{DNA methylation matrix (rows=CpG sites, columns=samples).}
}
\value{
Scores derived for each sample in \code{x}.
}
\description{
Horvath DNA methylation age estimate
}
\examples{
## Equivalent to the following:
x <- ... ## methylation matrix
standard <- meffonym.horvath.standard()
x.norm <- meffonym.bmiq.calibration(x, standard)
ret <- meffonym.score(x, "age.horvath")

}
