% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horvath.r
\name{meffonym.horvath.standard}
\alias{meffonym.horvath.standard}
\title{Horvath 21k methylation standard}
\usage{
meffonym.horvath.standard()
}
\description{
Obtained from
\url{http://labs.genetics.ucla.edu/horvath/dnamage/probeAnnotation21kdatMethUsed.csv}.
}
