% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile-normalization.r
\name{meffonym.quantile.normalization}
\alias{meffonym.quantile.normalization}
\title{Quantile normalization to a standard}
\usage{
meffonym.quantile.normalization(x, standard, ...)
}
\arguments{
\item{x}{DNA methylation matrix (rows=CpG sites, columns=samples).}

\item{standard}{A named vector of methylation levels corresponding to the row names of \code{x}.}
}
\value{
Matrix identical to \code{x} but with methylation levels
quantile normalized to the \code{standard}
and rows corresponding to \code{standard}.
Missing values in \code{x} are imputed by k-nearest neighbor.
}
\description{
Methylation levels are quantile normalized
so that they are comparable to a given dataset.
}
