% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miage.r
\name{miage}
\alias{miage}
\title{Estimate mitotic age from DNAm using the MiAge algorithm}
\usage{
miage(
  x,
  b = NULL,
  c = NULL,
  d = NULL,
  minage = 10,
  maxage = 10000,
  inits = c(500, minage + (1:4) * (maxage - minage)/5)
)
}
\arguments{
\item{x}{DNA methylation matrix (rows=CpG sites, columns=samples).
Missing values in \code{x} will be replaced with the mean value of the row.}

\item{b}{As defined above (Default: published values).}

\item{c}{As defined above (Default: published values).}

\item{d}{As defined above (Default: published values).}

\item{minage}{Lowest age to consider (Default: 10).}

\item{maxage}{Highest age to consider (Default: 10000).}

\item{inits}{Initial ages to consider at each iteration.
(Default: 500 and lowerage + (1:4)*(maxage-minage)/5).}
}
\value{
Estimated mitotic age for each sample.
}
\description{
Define the following:
* x[i,j] be the methylation level of CpG site i in sample j
* E(X[n,i]) be the expected methylation level of CpG site i after n generations,
* a[i] be the rate of de novo methylation of CpG site i
* b[i] be the fidelity of methylation maintenance of CpG site i
* c[i] = a[i]/(1-b[i]) 
* d[i] = E(X[1,i]) - c[i]
}
\details{
This function estimates the number of generations n[j] for each sample j
from the DNA methylation levels x[,j] of 268 CpG sites.
The estimate is obtained by selecting the value n[j] that minimizes the sum of
(E(X[n,i])-x[i,j])^2 across CpG sites i=1..268
where E(X[n,i]) is estimated as c[i] + b[i]^(n[j]-1)*d[i]
for given b[i], c[i] and d[i].
}
