% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo.r
\name{trafo}
\alias{trafo}
\title{Log-transform child ages}
\usage{
trafo(age, adult.age = 20)
}
\arguments{
\item{age}{Chronological age in years.}

\item{adult.age}{(Default: 20).}
}
\value{
Ages transformed so that
children are below zero and log-transformed
and adults are above zero.
}
\description{
Log-transform child ages
}
