% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.r
\name{meffonym.score}
\alias{meffonym.score}
\title{Calculate DNA methylation scores}
\usage{
meffonym.score(x, model)
}
\arguments{
\item{x}{DNA methylation matrix (rows=CpG sites, columns=samples).}

\item{model}{Model name from available list (\code{\link{meffonym.models}()}).}
}
\value{
Scores derived for each sample in \code{x}.
}
\description{
Calculate DNA methylation scores
}
